Successfully solved all errors and renamed registration app to teams
