+++ teams/urls.py
@@ -3,7 +3,7 @@ from django.urls import path
 from django.conf.urls import url
 
 
-app_name = 'registration'
+app_name = 'teams'
 
 urlpatterns = [
     path('ca', CampusAmbassadorRegisterView.as_view(), name='ca'),
