+++ templates/accounts/register.html
@@ -101,7 +101,7 @@
                         </button>
                     </div>
                     <div class="text-center p-t-136">
-                        <a class="txt2" href="{% url 'registration:ca' %}">
+                        <a class="txt2" href="{% url 'teams:ca' %}">
                             Register as Campus Ambassador
                             <i class="fa fa-long-arrow-right m-l-5" aria-hidden="true"></i>
                         </a>
