+++ templates/login.html
@@ -114,7 +114,7 @@
                         <i class="fa fa-long-arrow-right m-l-5" aria-hidden="true"></i>
                     </a>
                     <br>
-                    <a class="txt2" href="{% url 'registration:ca' %}">
+                    <a class="txt2" href="{% url 'teams:ca' %}">
                         Register as a Campus Ambassador
                         <i class="fa fa-long-arrow-right m-l-5" aria-hidden="true"></i>
                     </a>
