+++ templates/main/base.html
@@ -72,7 +72,7 @@
                                     <a href="{% url 'accounts:leaveTeam' %}">
                                         <li class="topbar-styles">Leave Team</li>
                                     </a>
-                                    <a href="{% url 'registration:remove_player' %}">
+                                    <a href="{% url 'teams:remove_player' %}">
                                         <li class="topbar-styles">Remove player</li>
                                     </a>
                                 </ul>
@@ -87,7 +87,7 @@
                                     <a href="{% url 'accounts:joinTeam' %}">
                                         <li class="topbar-styles">Join Team</li>
                                     </a>
-                                    <a href="{% url 'registration:team' %}">
+                                    <a href="{% url 'teams:team' %}">
                                         <li class="topbar-styles">Create Team</li>
                                     </a>
                                 </ul>
@@ -103,7 +103,7 @@
                                 <a href="{% url 'accounts:register' %}">
                                     <li class="topbar-styles">Register as Player</li>
                                 </a>
-                                <a href="{% url 'registration:ca' %}">
+                                <a href="{% url 'teams:ca' %}">
                                     <li class="topbar-styles">Campus Ambassador</li>
                                 </a>
                             </ul>
@@ -123,7 +123,7 @@
                                     <a href="{% url 'accounts:profile' %}">
                                         <li class="topbar-styles">Profile</li>
                                     </a>
-                                    <a href="{% url 'registration:ca' %}"
+                                    <a href="{% url 'teams:ca' %}"
                                        >
                                         <li class="topbar-styles">Campus Ambassador</li>
                                     </a>
@@ -185,7 +185,7 @@
                             {% else %}
                                 <li><a href="{% url 'accounts:joinTeam' %}">Join
                                     Team</a></li>
-                                <li><a href="{% url 'registration:team' %}"
+                                <li><a href="{% url 'teams:team' %}"
                                        >Create
                                     Team</a></li>
                             {% endif %}
@@ -202,7 +202,7 @@
                             <li><a href="{% url 'login' %}">Login</a></li>
                             <li><a herf="{% url 'accounts:register' %}">Register</a>
                             </li>
-                            <li><a href="{% url 'registration:ca' %}">CA</a></li>
+                            <li><a href="{% url 'teams:ca' %}">CA</a></li>
                         {% endif %}
                     </ul>
                 </nav>
