+++ templates/teams/ca_reg.html
