+++ templates/teams/password_reset_confirm.html
