+++ templates/teams/password_reset_done.html
