+++ templates/teams/password_reset_subject.txt
