+++ templates/teams/remove_player.html
